+++ app/configurator/components/chart-options-selector.tsx
@@ -58,14 +58,15 @@ import {
   isBarConfig,
   isColorInConfig,
   isComboChartConfig,
+  isMapConfig,
   isTableConfig,
   MapConfig,
   SortingType,
 } from "@/config-types";
 import {
+  getAxisDimension,
   getChartConfig,
   getMaybeValidChartConfigLimit,
-  getRelatedLimitDimension,
   useChartConfigFilters,
 } from "@/config-utils";
 import { ColorPalette } from "@/configurator/components/chart-controls/color-palette";
@@ -451,10 +452,11 @@ const EncodingOptionsPanel = (props: EncodingOptionsPanelProps) => {
 
   const hasSubOptions = encoding.options?.chartSubType ?? false;
 
-  const relatedLimitDimension = getRelatedLimitDimension({
-    chartConfig,
-    dimensions,
-  });
+  const limitMeasure = isMapConfig(chartConfig)
+    ? measures.find((m) => m.id === chartConfig.fields.symbolLayer?.measureId)
+    : isMeasure(component)
+      ? component
+      : undefined;
 
   return (
     <div
@@ -587,11 +589,11 @@ const EncodingOptionsPanel = (props: EncodingOptionsPanelProps) => {
           measures={measures}
         />
       )}
-      {isMeasure(component) && relatedLimitDimension ? (
+      {limitMeasure ? (
         <ChartLimits
           chartConfig={chartConfig}
-          measure={component}
-          relatedDimension={relatedLimitDimension}
+          dimensions={dimensions}
+          measure={limitMeasure}
         />
       ) : null}
       {encoding.options?.imputation?.shouldShow(chartConfig, observations) && (
@@ -760,39 +762,26 @@ const ChartLayoutOptions = ({
 
 const ChartLimits = ({
   chartConfig,
+  dimensions,
   measure,
-  relatedDimension,
 }: {
   chartConfig: ChartConfig;
+  dimensions: Dimension[];
   measure: Measure;
-  relatedDimension: Dimension;
 }) => {
   const [_, dispatch] = useConfiguratorState(isConfiguring);
   const filters = useChartConfigFilters(chartConfig);
   const onToggle = useEvent((checked: boolean, limit: Limit) => {
-    const { validRelated, wouldBeValid } = getMaybeValidChartConfigLimit({
-      chartConfig,
+    const actionProps = {
       measureId: measure.id,
-      limit,
-      relatedDimension,
-      filters,
-    });
-
-    if (!validRelated || !wouldBeValid) {
-      return;
-    }
-
-    const commonDispatchProps = {
-      measureId: measure.id,
-      relatedDimensionId: relatedDimension.id,
-      relatedDimensionValue: validRelated.dimensionValue,
+      related: limit.related,
     };
 
     if (checked) {
       dispatch({
         type: "LIMIT_SET",
         value: {
-          ...commonDispatchProps,
+          ...actionProps,
           color: "#ff0000",
           lineType: "solid",
         },
@@ -800,10 +789,11 @@ const ChartLimits = ({
     } else {
       dispatch({
         type: "LIMIT_REMOVE",
-        value: commonDispatchProps,
+        value: actionProps,
       });
     }
   });
+  const axisDimension = getAxisDimension({ chartConfig, dimensions });
 
   const availableLimitOptions = useMemo(() => {
     return measure.limits
@@ -812,8 +802,8 @@ const ChartLimits = ({
           getMaybeValidChartConfigLimit({
             chartConfig,
             measureId: measure.id,
+            axisDimension,
             limit,
-            relatedDimension: relatedDimension,
             filters,
           });
 
@@ -831,10 +821,10 @@ const ChartLimits = ({
         };
       })
       .filter(truthy);
-  }, [chartConfig, relatedDimension, filters, measure.id, measure.limits]);
+  }, [axisDimension, chartConfig, filters, measure.id, measure.limits]);
 
   const { data: userPalettes } = useUserPalettes();
-  const paletteId = get(chartConfig, `fields.color.paletteId`);
+  const paletteId = get(chartConfig, "fields.color.paletteId");
   const colors = getPalette({
     paletteId,
     fallbackPalette: userPalettes?.find((d) => d.paletteId === paletteId)
@@ -896,8 +886,7 @@ const ChartLimits = ({
                           type: "LIMIT_SET",
                           value: {
                             measureId: measure.id,
-                            relatedDimensionId: relatedDimension.id,
-                            relatedDimensionValue: maybeLimit!.dimensionValue,
+                            related: limit.related,
                             color,
                             lineType,
                           },
@@ -924,8 +913,7 @@ const ChartLimits = ({
                           type: "LIMIT_SET",
                           value: {
                             measureId: measure.id,
-                            relatedDimensionId: relatedDimension.id,
-                            relatedDimensionValue: maybeLimit!.dimensionValue,
+                            related: limit.related,
                             color,
                             lineType: "solid",
                           },
@@ -946,8 +934,7 @@ const ChartLimits = ({
                           type: "LIMIT_SET",
                           value: {
                             measureId: measure.id,
-                            relatedDimensionId: relatedDimension.id,
-                            relatedDimensionValue: maybeLimit!.dimensionValue,
+                            related: limit.related,
                             color,
                             lineType: "dashed",
                           },